feat: Sidebar가 DOM구조를 반환하도록 변경
